+++ pkg/analyzer/lib/dart/element/type.dart
@@ -133,6 +133,9 @@ abstract class DartType {
   ///
   /// Note too that the current implementation of this method is only guaranteed
   /// to work when the parameter types are type variables.
+  @Deprecated("""
+Use ClassElement.instantiate() or FunctionTypeAliasElement.instantiate()
+""")
   DartType substitute2(
       List<DartType> argumentTypes, List<DartType> parameterTypes);
 }
@@ -201,6 +204,7 @@ abstract class FunctionType implements ParameterizedType {
   @override
   FunctionType instantiate(List<DartType> argumentTypes);
 
+  @Deprecated("Use FunctionTypeAliasElement.instantiate()")
   @override
   FunctionType substitute2(
       List<DartType> argumentTypes, List<DartType> parameterTypes);
@@ -433,6 +437,7 @@ abstract class InterfaceType implements ParameterizedType {
   PropertyAccessorElement lookUpSetterInSuperclass(
       String name, LibraryElement library);
 
+  @Deprecated("Use ClassElement.instantiate()")
   @override
   InterfaceType substitute2(
       List<DartType> argumentTypes, List<DartType> parameterTypes);